# qnop-api/src/main/resources/openapi/openapi.yaml
...
openapi: 3.1.0
info:
  title: qnop API
  version: v1
  description: |
    **qnop** ("Qualified Notes on Papers") is an enterprise document-review system.
    This is the published, versioned REST contract consumed by the qnop frontend
    and by third-party integrators. It carries `/api/v1` URL versioning and a
    deprecation policy (ADR-0015).
  license:
    name: AGPL-3.0-only
    url: https://www.gnu.org/licenses/agpl-3.0.html
servers:
  - url: /api/v1
    description: Versioned API root
...
paths:
...
  /admin/users:
    get:
      tags:
        - AdminUsers
      summary: List and search users
      description: |
        Returns a page of users (internal and external), most-recent first.
        Optional `q` matches display name, email or username (case-insensitive);
        optional `role` filters by global role.
      operationId: listUsers
      security:
        - bearerAuth: []
      parameters:
        - name: q
          in: query
          required: false
          description: Case-insensitive search over display name, email and username.
          schema:
            type: string
            maxLength: 200
        - name: role
          in: query
          required: false
          description: Filter by global role.
          schema:
            $ref: '#/components/schemas/UserRole'
        - name: enabled
          in: query
          required: false
          description: Filter by account status; omit for all.
          schema:
            type: boolean
        - name: sort
          in: query
          required: false
          description: |
            Sort as `field,direction`. Allowed fields: displayName, email,
            username, role, createdAt, lastLoginAt; direction asc|desc.
          schema:
            type: string
            default: displayName,asc
        - name: page
          in: query
          required: false
          description: Zero-based page index.
          schema:
            type: integer
            format: int32
            minimum: 0
            default: 0
        - name: size
          in: query
          required: false
          description: Page size.
          schema:
            type: integer
            format: int32
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: A page of users
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminUserListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      tags:
        - AdminUsers
      summary: Create or invite a user
      description: |
        Creates an internal account. If `initialPassword` is provided the account
        is enabled immediately and must change its password on first login. If it
        is omitted, the account is created without a password and an invitation
        link (set-your-password) is emailed to the address.
      operationId: createUser
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminUserCreateRequest'
      responses:
        '201':
          description: The created user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminUserDetail'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: A user with that username or email already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /admin/users/{userId}:
    get:
      tags:
        - AdminUsers
      summary: Get a user
      operationId: getUser
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/UserIdParam'
      responses:
        '200':
          description: The user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminUserDetail'
        '404':
          description: Unknown user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    patch:
      tags:
        - AdminUsers
      summary: Update a user
      description: |
        Partial update of display name, global role and enabled state. The
        self-lockout guard rejects (409) an attempt to disable or demote your own
        account or the last remaining admin.
      operationId: updateUser
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/UserIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminUserUpdateRequest'
      responses:
        '200':
          description: The updated user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminUserDetail'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Unknown user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Self-lockout or last-admin guard violated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    delete:
      tags:
        - AdminUsers
      summary: Delete a user
      description: |
        Permanently deletes the user and, by cascade, their tokens, settings,
        OIDC identity and team memberships. The self-lockout guard rejects (409)
        deleting your own account or the last enabled admin.
      operationId: deleteUser
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/UserIdParam'
      responses:
        '204':
          description: User deleted
        '404':
          description: Unknown user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Self-lockout or last-admin guard violated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /admin/users/{userId}/password-reset:
    post:
      tags:
        - AdminUsers
      summary: Reset a user's password (admin)
      description: |
        Revokes the user's active sessions immediately, issues a single-use
        password-reset token and emails the set-your-password link. Works for any
        internal account (e.g. to (re)send an invitation or to help a locked-out
        user). External (OIDC) accounts have no local password and are rejected
        with 409. When email could not be sent, the response carries `resetUrl`
        so the operator can deliver it out of band.
      operationId: sendUserPasswordReset
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/UserIdParam'
      responses:
        '200':
          description: The reset was triggered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminPasswordResetResponse'
        '404':
          description: Unknown user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: The account has no local password (external/OIDC)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /admin/users/{userId}/password:
    post:
      tags:
        - AdminUsers
      summary: Generate a new password for a user (admin)
      description: |
        Sets a freshly generated strong password on the user, marks it for change
        on first login, and revokes the user's active sessions immediately. The
        generated password is returned **once** for the admin to hand over — it is
        never stored in clear text and cannot be retrieved again. External (OIDC)
        accounts have no local password and are rejected with 409. An admin may
        target their own account; doing so ends their sessions (the response still
        carries the new password to sign back in with).
      operationId: generateUserPassword
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/UserIdParam'
      responses:
        '200':
          description: The password was generated and set
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminGeneratedPasswordResponse'
        '404':
          description: Unknown user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: The account has no local password (external/OIDC)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT access token issued by the qnop auth subsystem (epic #7). Sent as
        `Authorization: Bearer <token>` on authenticated endpoints. The auth flow
        itself is introduced in a later issue; the scheme is declared here so the
        contract is complete from the first endpoint.
  parameters:
...
    UserIdParam:
      name: userId
      in: path
      required: true
      description: The user's id.
      schema:
        type: string
        format: uuid
...
  responses:
    Unauthorized:
      description: Authentication is required (uniform error envelope, issue #45).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: The caller lacks permission for this resource (issue #45).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
...
    UserRole:
      type: string
      description: |
        A user's global system role — exactly one per user (issue #98). `ADMIN`
        administers the system; `MEMBER` is the standard user; `AUDITOR` has
        organisation-wide read access to the audit trail and compliance.
      enum:
        - ADMIN
        - MEMBER
        - AUDITOR
    UserSource:
      type: string
      description: |
        How the account authenticates: `INTERNAL` (local username + password) or
        `EXTERNAL` (provisioned from an OIDC/OAuth2 provider).
      enum:
        - INTERNAL
        - EXTERNAL
...
    AdminUserSummary:
      type: object
      description: A user as shown in the admin user list.
      properties:
        id:
          type: string
          format: uuid
        displayName:
          type: string
        email:
          type: string
        username:
          type: string
          nullable: true
          description: The local username; null for external (OIDC) accounts.
        role:
          $ref: '#/components/schemas/UserRole'
        source:
          $ref: '#/components/schemas/UserSource'
        enabled:
          type: boolean
        passwordChangeRequired:
          type: boolean
          description: Whether the user must change their password before normal use.
        providerName:
          type: string
          nullable: true
          description: The OIDC provider's name for external accounts; null otherwise.
        lastLoginAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        avatarUrl:
          type: string
          nullable: true
          description: |
            URL of the user's profile picture (with a cache-busting `?v=` token),
            or null when none is set — the UI then renders the initials avatar.
      required:
        - id
        - displayName
        - email
        - role
        - source
        - enabled
        - passwordChangeRequired
        - createdAt
    AdminUserDetail:
      allOf:
        - $ref: '#/components/schemas/AdminUserSummary'
      description: Full detail for a single user (currently the same fields as the summary).
    AdminPasswordResetResponse:
      type: object
      description: Outcome of an admin-triggered password reset.
      properties:
        emailSent:
          type: boolean
          description: Whether the reset email was delivered.
        resetUrl:
          type: string
          nullable: true
          description: |
            The single-use reset link, returned only when the email could not be
            sent so the operator can deliver it out of band.
      required:
        - emailSent
    AdminGeneratedPasswordResponse:
      type: object
      description: |
        A freshly generated password, returned exactly once. It is never stored in
        clear text and cannot be retrieved again — the admin must hand it over now.
      properties:
        password:
          type: string
          description: The generated plaintext password, shown a single time.
      required:
        - password
    AdminUserListResponse:
      type: object
      description: A page of users for the admin list.
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/AdminUserSummary'
        total:
          type: integer
          format: int64
          description: Total number of users matching the query.
        page:
          type: integer
          format: int32
          description: Zero-based page index.
        size:
          type: integer
          format: int32
          description: Page size.
      required:
        - items
        - total
        - page
        - size
    AdminUserCreateRequest:
      type: object
      description: |
        Create an internal user. Provide `initialPassword` to enable the account
        immediately (with a forced change on first login), or omit it to send an
        invitation link instead.
      properties:
        displayName:
          type: string
          minLength: 1
          maxLength: 200
        username:
          type: string
          minLength: 3
          maxLength: 64
        email:
          type: string
          format: email
          maxLength: 320
        role:
          $ref: '#/components/schemas/UserRole'
        initialPassword:
          type: string
          minLength: 8
          maxLength: 200
          nullable: true
          description: Optional. When omitted, an invitation email is sent instead.
      required:
        - displayName
        - username
        - email
        - role
    AdminUserUpdateRequest:
      type: object
      description: Partial update; only provided fields change.
      properties:
        displayName:
          type: string
          minLength: 1
          maxLength: 200
          nullable: true
        role:
          allOf:
            - $ref: '#/components/schemas/UserRole'
          nullable: true
        enabled:
          type: boolean
          nullable: true
...
    ErrorResponse:
      type: object
      description: |
        Standard error envelope returned for non-2xx responses across the API.
        Carries a stable machine-readable `code` and a human-readable `message`;
        sensitive internals are never leaked (java/security rule).
      properties:
        code:
          type: string
          description: Stable, machine-readable error code (e.g. `NOT_FOUND`).
        message:
          type: string
          description: Human-readable, non-sensitive error description.
        timestamp:
          type: string
          format: date-time
          description: Server timestamp (RFC 3339) when the error was produced.
      required:
        - code
        - message
        - timestamp